# api/docs/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/verify/check": {
      "post": {
        "operationId": "Verify_CheckConfirmCode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authCheckConfirmCodeResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/authCheckConfirmCodeReq"
            }
          }
        ],
        "tags": [
          "Verify"
        ]
      }
    },
    "/v1/verify/email": {
      "post": {
        "operationId": "Verify_VerifyEmail",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Verify"
        ]
      }
    },
    "/v1/verify/send": {
      "post": {
        "operationId": "Verify_SendConfirmCode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authInquiryResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Verify"
        ]
      }
    }
  },
  "definitions": {
...
    "authCheckConfirmCodeReq": {
      "type": "object",
      "properties": {
        "eve": {
          "$ref": "#/definitions/authevents"
        },
        "code": {
          "type": "string"
        }
      }
    },
    "authCheckConfirmCodeResp": {
      "type": "object",
      "properties": {
        "eve": {
          "$ref": "#/definitions/authevents"
        }
      }
    },
    "authEmpty": {
      "type": "object"
    },
    "authInquiryResp": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        }
      }
    },
...
    "authevents": {
      "type": "string",
      "enum": [
        "LOGIN",
        "CHANGE_PASSWORD",
        "CHANGE_USERNAME",
        "CHANGE_EMAIL",
        "CONFIRM_EMAIL"
      ],
      "default": "LOGIN"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}